moved transactions state out to parent component of transactions component
