+++ src/Transactions.js
@@ -37,52 +37,7 @@ const Transactions = () => {
         }
         
     ];
-
-    const formatter = new Intl.NumberFormat('en-US', {
-        style: 'currency',
-        currency: 'USD',
-        minimumFractionDigits: 2
-    })
-
-    const getTransactions = () => {
-        dispatch(setLoading(true))
-        let i = 1
-        let transactionRows = []
-        accounts.forEach((account) => {
-            fetch("http://localhost:3000/transactions",{
-                method: "POST",
-                headers: {"Content-Type": "application/json"},
-                body: JSON.stringify({plaid_account_id: account.id})
-            })
-            .then(res => res.json())
-            .then(data => {
-                if(data.auth){
-                    data.transactions.forEach((transaction) => {
-                        fetch(`http://localhost:3000/transaction_categories/${transaction.transaction_category_id}`)
-                        .then(resp => resp.json())
-                        .then(categoryData => {
-                            if(categoryData.auth){
-                                if(transaction.value < 0){
-                                    transactionRows.push({id: (i++), Date: transaction.date, Description: transaction.description, Category: categoryData.transaction_category.name, Value: formatter.format(transaction.value), Bank: account.p_institution, Type: "Expense"})
-                                }
-                                else{
-                                    transactionRows.push({id: (i++), Date: transaction.date, Description: transaction.description, Category: categoryData.transaction_category.name, Value: formatter.format(transaction.value), Bank: account.p_institution, Type: "Income"})
-                                }
-                            }
-                            else{
-                                alert(categoryData.message)
-                            }
-                        })
-                    })
-                }
-                else{
-                    alert(data.message)
-                }
-            })
-        })
-        dispatch(setTransactions(transactionRows))
-        dispatch(setLoading(false))
-    }
+    
 
     Array.prototype.remove = function() {
         var what, a = arguments, L = a.length, ax;
@@ -122,7 +77,6 @@ const Transactions = () => {
     }
 
     useEffect(() => {
-        getTransactions()
     }, [])
 
     return (
